Add uiManager.js module for improved modularity
- Extract UI management functions to separate module
- Implement ES6 module imports/exports
- Update main JS file to import from uiManager.js
- Refactor UI event listeners and state restoration
- Ensure proper integration with other modules
